create an Interface to generate both Marmaid and Plant. verify CheckStyle
