+++ src/main/java/com/github/veluvexiau/umldoc/core/plantExport.md
@@ -130,7 +130,16 @@
 	class Dependency{
 	}
 
+	class PlantUMLExtract{
+	}
+
+	class MermaidExtract{
+	}
+
 	class Main{
+	}
+
+	class Extract{
 	}
 
 	class Main_1{
